correos-backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# SOAP Client
zeep==4.2.1
lxml==4.9.3
requests==2.31.0

# PDF generation
fpdf2==2.8.2
